Move code to the parser
